Split admin api anon proxy files
They are too large to keep in a single anon_proxy.rb file,
make them top-level admin resource endpoints like everything else.
